docker_crud/requirements.txt
Django
django-filter
djangorestframework
